BGDIINF_SB-2890 : PR review
Do not set the same projection (and do reproject processes)
renaming of the plugin that now only reproject selected features
